+++ rhi/dx9/swapchain_dx9.cpp
@@ -42,7 +42,7 @@ void CDx9RhiSwapChain::GetBuffers(CVector<IRhiRenderTarget*>& buffers)
 void CDx9RhiSwapChain::ResizeBuffers()
 {
 	auto& presentParams = m_parent->m_presentParams;
-	presentParams.BackBufferCount = m_bufferCount;
+	presentParams.BackBufferCount = (UINT)m_bufferCount;
 	auto videoSystem = m_parent->m_parent->m_videoSystem;
 	presentParams.BackBufferWidth = videoSystem->GetWidth();
 	presentParams.BackBufferHeight = videoSystem->GetHeight();
@@ -73,7 +73,7 @@ void CDx9RhiSwapChain::ResizeBuffers()
 	for (ssize i = 0; i < m_bufferCount; i++)
 	{
 		IDirect3DSurface9* surface = nullptr;
-		result = m_handle->GetBackBuffer(i, D3DBACKBUFFER_TYPE_MONO, &surface);
+		result = m_handle->GetBackBuffer((UINT)i, D3DBACKBUFFER_TYPE_MONO, &surface);
 		if (FAILED(result))
 		{
 			if (m_buffers.Size() < 1)
